chore(packages/shared): relocate isErrorReport to objectValidators
